% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNumericRange.R
\name{checkNumericRange}
\alias{checkNumericRange}
\title{Compares each value in passed data.frame to specified minimum and maximum values.
Subject IDs and associated column names are printed out for values that are not
within specified range.}
\usage{
checkNumericRange(dat_cols, min, max)
}
\arguments{
\item{dat_cols}{data.frame with columns of items to be checked for specified range}

\item{min}{minimum value in range}

\item{max}{maximum value in range}
}
\description{
Compares each value in passed data.frame to specified minimum and maximum values.
Subject IDs and associated column names are printed out for values that are not
within specified range.
}
