% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearestRow.R
\name{getNearestRow}
\alias{getNearestRow}
\title{Get analytic site visit nearest MRI visit}
\usage{
getNearestRow(x, multivis)
}
\arguments{
\item{x}{Row of REDCap data to use to identify analytic site visit.}

\item{multivis}{The data frame containing the analytic site information.}
}
\description{
Takes a row of the REDCap data and compares it to the multivis data that
contains infromation for each subject's analytic site visit.
}
