% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logError.R
\name{logError}
\alias{logError}
\title{Log errors in data frame.}
\usage{
logError(dat, column, problem, variable, save.file = NA,
  warning = FALSE, id.column = "idnum")
}
\arguments{
\item{dat}{Data file containing an ID column.}

\item{column}{A logical vector the same length as dat.}

\item{problem}{A short string describing the problem at the TRUE values
of the string.}

\item{save.file}{A string describing the location of the file to save to.
Always appends.}

\item{id.number}{The name of the column that contains the ID values.
Defaults to "idnum".}
}
\description{
Takes a data frame and logs the idnums of subjects who fit a given criteria
and logs them to a given file.
}
\examples{

Take dat$age: [70, 88, 71, NA]

> logError(dat, is.na(dat$age), "Age value is NA", "errors.txt")

errors.txt would read something like:

0004 Age value is NA

}
