% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericColumnsUPDRS.R
\name{numericColumnsUPDRS}
\alias{numericColumnsUPDRS}
\title{Returns modified version of passed data.frame with UPDRS elements converted to numeric.}
\usage{
numericColumnsUPDRS(dat)
}
\arguments{
\item{dat}{data.frame to change UPDRS columns to numeric}
}
\description{
Returns modified version of passed data.frame with UPDRS elements converted to numeric.
}
