% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameFreeSurfer.R
\name{renameFreeSurfer}
\alias{renameFreeSurfer}
\title{Score Freezing of Gait Instrument}
\usage{
renameFreeSurfer(dat)
}
\arguments{
\item{dat}{Input data frame with freesurfer variables}
}
\value{
dat New data frame with freesurfer variables renamed
}
\description{
Score Freezing of Gait Instrument
}
