% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreFOG.R
\name{scoreFOG}
\alias{scoreFOG}
\title{Score Freezing of Gait Instrument}
\usage{
scoreFOG(dat)
}
\arguments{
\item{dat}{Input data frame with fog variables}
}
\value{
dat New data frame with FOG score added to it
}
\description{
Score Freezing of Gait Instrument
}

