% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSAI.R
\name{scoreSAI}
\alias{scoreSAI}
\title{Score SAI}
\usage{
scoreSAI(dat)
}
\arguments{
\item{dat}{Input data frame with SAI variables}
}
\value{
dat New data frame with SAI score added to it
}
\description{
Score SAI
}
