% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreUPDRS.R
\name{scoreUPDRS}
\alias{scoreUPDRS}
\title{scoreUPDRS; Modifies original data.frame by appending total UPDRS columns for UPDRS Part 3 & 4.}
\usage{
scoreUPDRS(dat)
}
\arguments{
\item{dat}{Input data frame with items to be scored.}
}
\value{
dat Output data frame with items scored.
}
\description{
scoreUPDRS; Modifies original data.frame by appending total UPDRS columns for UPDRS Part 3 & 4.
}
