% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testVariable.R
\name{testVariable}
\alias{testVariable}
\title{Test demographic variable}
\usage{
testVariable(col, all.subjects)
}
\arguments{
\item{col}{Column name}

\item{all.subjects}{Data frame containing information for all subjects.}
}
\value{
Vector of length 7 with PD group mean/count; PD group SD/percent;
control group mean/count; control group SD/percent; p-value (when
appropriate); total mean/count; total SD/percent
}
\description{
Given a column and data frame, compares values of column for PD and
control/unaffected participants.
}
