% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformGnB.R
\name{transformGnB}
\alias{transformGnB}
\title{Transform gait and balance data}
\usage{
transformGnB(dat)
}
\arguments{
\item{dat}{A data frame with the gait and balance data}
}
\description{
Cleans up column names, and process gait and balance data for upload to
REDCap.
}
