% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udallCalculateCost.R
\name{udallCalculateCost}
\alias{udallCalculateCost}
\title{Calculate task cost for gait and balance}
\usage{
udallCalculateCost(data.frame, st.cols, dt.cols)
}
\arguments{
\item{data.frame}{Data frame that contains both the single task and dual
task cost columsn.}

\item{st.cols}{Which columns have information on single-task performance.}

\item{dt.cols}{Which columns have information on dual-task performance.
Length must match st.cols.}
}
\value{
Data frame with #cols = length of st.cols and
nrows = nrows(data.frame). Cost measurements for
}
\description{
Calculate task cost for gait and balance
}
