% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udallCleanREDCapDataWide.R
\name{udallCleanREDCapDataWide}
\alias{udallCleanREDCapDataWide}
\title{Clean up the data that come from redcap Udall}
\usage{
udallCleanREDCapDataWide(dat)
}
\arguments{
\item{dat}{Data frame that comes directly from redcap_read}
}
\value{
A cleaned data frame, with NA and empty columns removed, on or off prepended to the names, and all subject data on a single row (wide format)
}
\description{
Clean up the data that come from redcap Udall
}

