% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udallCrosscheckEntries.R
\name{udallCrosscheckEntries}
\alias{udallCrosscheckEntries}
\title{Crosscheck entries}
\usage{
udallCrosscheckEntries(cdat, error.file)
}
\arguments{
\item{cdat}{Cleaned data frame (producet of udallCleanDataWide)}

\item{error.file}{File to save all errors to.}
}
\description{
Crosschecks the double-uploaded data and between the REDCap data and the
MS Access/Analytic Core/mulitvis data.
}
\examples{
udallCrosscheckEntries(cdat, "mismatch-errors.txt")

}
