% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udallGetAnalyze.R
\name{udallGetAnalyze}
\alias{udallGetAnalyze}
\title{Extracts which data collection measures are to be included in the analysis}
\usage{
udallGetAnalyze(dat, whys = FALSE)
}
\arguments{
\item{whys}{Do you need the explanations? Default FALSE}

\item{cdat}{A cleaned data frame}
}
\value{
A data frame with the analyze values, or a list of two data frames
with the analyze values and the corresponding whys
}
\description{
udallGetAnalyze returns a dataframe of dimensions n x 9, where n is the
number of subjects. The first column contains idnums and the last eight
logical values. If the `whys` option is selected, the function will return
a list of two data frames, where the second has the explanations downloaded
frome REDCap.
}
