% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udallProcessGnB.R
\name{udallProcessGnB}
\alias{udallProcessGnB}
\title{Process gait and balance data}
\usage{
udallProcessGnB(xlsx, write = NA)
}
\arguments{
\item{xlsx}{Location of the xlsx file with the relevant gait and balance
data, relative to the current working directory.}

\item{write}{Location to save output CSV files to. Defaults to NA, which
signals no write-out.}
}
\value{
data.l A list of data frames with the transformed data.
}
\description{
Process gait and balance data
}
