% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udallReplaceMissing.R
\name{udallReplaceMissing}
\alias{udallReplaceMissing}
\title{Replace missing data codes with NA for analysis}
\usage{
udallReplaceMissing(dat)
}
\arguments{
\item{dat}{Data frame with missing data codes}
}
\value{
dat Data frame with missing data codes assigned to NA
}
\description{
Replace missing data codes with NA for analysis
}

