% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warningIfNot.R
\name{warningIfNot}
\alias{warningIfNot}
\title{Give a warning if the condition is not true, consisting of message}
\usage{
warningIfNot(condition, message)
}
\arguments{
\item{condition}{An expression or flag that evaluates to a boolean}

\item{message}{A message to be printed if condition is FALSE (not true)}
}
\description{
Give a warning if the condition is not true, consisting of message
}

