% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udallCalculateCost.R
\name{udallCalculateCost}
\alias{udallCalculateCost}
\title{Calculate task cost for gait and balance}
\usage{
udallCalculateCost(single.task, dual.task)
}
\arguments{
\item{single.task}{Data from single task run, as a data frame.}

\item{dual.task}{Data from dual task run, as a data.frame}
}
\value{
Cost for all measures over all subjects
}
\description{
Calculate task cost for gait and balance
}
