% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udallDemographicTable.R
\name{udallDemographicTable}
\alias{udallDemographicTable}
\title{Create a demographic table for Udall P2}
\usage{
udallDemographicTable(data, columns, group.by, group.ok = NULL)
}
\arguments{
\item{data}{Data from which to select the columns.}

\item{columns}{List of columns to summarize, all must exist in data.}

\item{group.by}{Variable to group subjects by.}

\item{group.ok}{List of groups to include in the sorting; defaults to all
options in group.by if not given.}
}
\value{
A data frame with the inputted columns on the left and seven columns:
PD group mean/count; PD group SD/percent; control group mean/count; control
group SD/percent; p-value (when appropriate); total mean/count; total
SD/percent
}
\description{
Creates a demographic table from the given data frame and columns. Make sure
to type columns appropriately before inputting them into
udallDemographicTable (e.g. factor conversions).
}
